* Put parentheses around casts consisting of two or more words (fixes compilation on GCC) * Fixed a few int-to-float cast warnings.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1376 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Game/GameSetup.cpp
@@ -162,7 +162,7 @@ bool CGameSetup::Init(char* buf, int size)
 		float3 defaultCol(palette.teamColor[colorNum][0]/255.0,palette.teamColor[colorNum][1]/255.0,palette.teamColor[colorNum][2]/255.0);
 		float3 color=file.GetFloat3(defaultCol,s+"rgbcolor");
 		for(int b=0;b<3;++b)
-			gs->Team(a)->color[b]=color[b]*255;
+			gs->Team(a)->color[b]=int(color[b]*255);
 
  		gs->Team(a)->handicap=atof(file.SGetValueDef("0",s+"handicap").c_str())/100+1;
  		gs->Team(a)->leader=atoi(file.SGetValueDef("0",s+"teamleader").c_str());
